Commit pendente

+++ routes/tasks.js
@@ -2,11 +2,15 @@ module.exports = app => {
     const Tasks = app.db.models.Tasks;
 
     app.route("/tasks")
+    .all(app.auth.authenticate())
     .get((req, res) => {
-        Tasks.findAll({}).then(result => res.json(result))
+        Tasks.findAll({
+            where: {user_id: req.user.id }
+        }).then(result => res.json(result))
         .catch(error => {res.status(412).json({msg: error.message});});
     })
     .post((req, res) => {
+        req.body.user_id = req.user.id;
         Tasks.create(req.body)
         .then(result => res.json(result))
         .catch(error => {
@@ -15,9 +19,10 @@ module.exports = app => {
     });
 
     app.route("/tasks/:id")
+    .all(app.auth.authenticate())
     .get((req, res) => {
         // "/tasks/1": Consulta uma tarefa
-        Tasks.findOne({where: req.params})
+        Tasks.findOne({where: {id: req.params.id, user_id: req.user.id }})
             .then(result => {
                 if (result) {
                     res.json({result});
@@ -31,7 +36,7 @@ module.exports = app => {
     })
     .put((req, res) =>{
         // "/tasks/1": Atualiza uma tarefa
-        Tasks.update(req.body, {where: req.params})
+        Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(402).json({msg: error.message});
@@ -39,7 +44,7 @@ module.exports = app => {
     })
     .delete((req, res) => {
         // "/tasks/1": Exclui uma tarefa uma tarefa
-        Tasks.destroy({where: req.params})
+        Tasks.destroy({where: { id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(412).json({msg: error.message});
